+++ test/test_helper.rb
@@ -22,14 +22,5 @@ I18n.load_path << File.expand_path("../locales/en.yml", __FILE__)
 
 class Test::Unit::TestCase
   include Vagrant::TestHelpers
-
-  # Sets up the mocks and stubs for a downloader
-  def mock_downloader(downloader_klass)
-    tempfile = mock("tempfile")
-    tempfile.stubs(:write)
-
-    _, env = action_env
-    [downloader_klass.new(env), tempfile]
-  end
 end