+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.6.4 (unreleased)
 
+  - Default VM name is now properly the parent folder of the working directory
+    of the environment.
   - Added method to `TestHelpers` to assist with testing new downloaders.
   - `up --no-provision` works again. This disables provisioning during the
     boot process.